requirements.txt
aiogram>=3.6.0
SQLAlchemy>=2.0.30
aiosqlite>=0.20.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
apscheduler>=3.10.4
pytest>=8.2.0
httpx>=0.27.0

